requirements.txt
mido==1.2.10
pydub==0.25.1
python-rtmidi==1.4.9
PyYAML==6.0
pyzmq==25.0.2
